backend/requirements.txt
# FastAPI Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Temporal SDK
temporalio==1.5.1

# Payment Processing
stripe==8.2.0

# AWS SDK
boto3==1.34.50

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Error Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# Security & Validation
bleach==6.1.0

# System Monitoring
psutil==5.9.8

# Rate Limiting (Production)
redis==5.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Development
black==24.2.0
ruff==0.2.2
mypy==1.8.0
